feat: sync timer attached to main editor page
